+++ ChangeLog.md
@@ -10,10 +10,10 @@ All notable changes to this project will be documented in this file.
 - NEW : Add a button to the commercial proposal form that allows you to create a travel plans *25/04/2024* - 1.7.0
 
 ## 1.6
-- FIX : Print default price when you create a travel plan  *25/04/2024* - 1.6.0
+- FIX : Print default price when you create a travel plan - *25/04/2024* - 1.6.0
 
 ## 1.5
-- NEW : Set default price for travel plan  *19/04/2024* - 1.4.0
+- NEW : Set default price for travel plan  *24/04/2024* - 1.5.0
 
 ## 1.4